Merging my register code into Yanuar's main program.
I have edited Yanuar’s main program and merged it with my register code
so that its no longer the dummy register code. Now it is connected with
the database, therefore if the npm/username is the same they cannot use
that npm/username to register. I have also made validation for if the
user entered an empty field, the program will keep asking until the
user enters the proper data.